Move the handling of function call errors to the adapter
The idea is that the adater will be able to decide not to raise an error and instead continue the evaluation with, for example, null as result.

This removes the ArityException throwing branch from the type runtime function call evaluation – this should no longer happen during runtime since arities are checked at compile time. If an arity error is detected during runtime an IllegalStateException will be thrown instead.